handle category, comment edge cases
Restrict Category creation to logged in users only, and Post editing to
the creator of the posts. Add an if clause to the _comment template so
that it doesn't blow up if the comment.id is nil (in the case of a
comment that is created but not saved in the database).